--- src/main/java/supernova/whokie/user/controller/dto/Profile.java
@@ -1,5 +1,0 @@
-package supernova.whokie.user.controller.dto;
-
-public record Profile(String profile_image_url) {
-
-}
