+++ src/main/java/supernova/whokie/user/infrastructure/apiCaller/dto/Profile.java
@@ -1,0 +1,5 @@
+package supernova.whokie.user.infrastructure.apiCaller.dto;
+
+public record Profile(String profile_image_url) {
+
+}
