+++ src/main/java/supernova/whokie/user/infrastructure/apiCaller/dto/UserInfoResponse.java
@@ -1,4 +1,4 @@
-package supernova.whokie.user.controller.dto;
+package supernova.whokie.user.infrastructure.apiCaller.dto;
 
 public record UserInfoResponse(Long id, KakaoAccount kakao_account) {
